requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
backoff==2.2.1
certifi==2024.6.2
charset-normalizer==3.3.2
colorlog==6.8.2
datasets==2.14.7
dill==0.3.7
distro==1.9.0
dspy-ai==2.4.9
filelock==3.15.1
frozenlist==1.4.1
fsspec==2023.10.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.3
idna==3.7
joblib==1.3.2
Mako==1.3.5
MarkupSafe==2.1.5
multidict==6.0.5
multiprocess==0.70.15
numpy==1.26.4
openai==1.34.0
optuna==3.6.1
packaging==24.1
pandas==2.2.2
pyarrow==16.1.0
pyarrow-hotfix==0.6
pydantic==2.7.4
pydantic_core==2.18.4
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
structlog==24.2.0
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.1
xxhash==3.4.1
yarl==1.9.4
